Change: move VTs into target, sensor into own file
To be easier to reference within the documentation the VTs are moved
into target since it is just used there.

Since the sensor doesn't have a direct connection to a scan it is moved
to an own file description within the same package.
