+++ c/example/generate_description.c
@@ -140,10 +140,13 @@ struct TOC {
 struct AggregateTOC *build_aggregate_toc(enum eulabeia_aggregate aggregate)
 {
 	struct AggregateTOC *toc;
-	fprintf(stderr, "%s: build toc for %s\n", __func__, eulabeia_aggregate_to_str(aggregate));
+	fprintf(stderr,
+		"%s: build toc for %s\n",
+		__func__,
+		eulabeia_aggregate_to_str(aggregate));
 	toc = calloc(1, sizeof(*toc));
 #define X(a, b, c)                                                             \
-	if (is_relevant(aggregate, a)) {                                     \
+	if (is_relevant(aggregate, a)) {                                       \
 		if (toc->len == toc->cap) {                                    \
 			toc->cap += 10;                                        \
 			if ((toc->links = realloc(                             \
@@ -156,8 +159,8 @@ struct AggregateTOC *build_aggregate_toc(enum eulabeia_aggregate aggregate)
 				exit(23);                                      \
 		}                                                              \
 		toc->links[toc->len] =                                         \
-		    message_type_aggregate_link(a, aggregate);               \
-		toc->types[toc->len] = a;                                    \
+		    message_type_aggregate_link(a, aggregate);                 \
+		toc->types[toc->len] = a;                                      \
 		toc->len++;                                                    \
 	}
 	EULABEIA_MESSAGE_TYPES
@@ -250,7 +253,10 @@ static void print_entry(enum eulabeia_aggregate aggregate,
 	char *id, *example, *response, *addition = NULL;
 	enum eulabeia_message_type type;
 	int i, modify;
-	fprintf(stderr, "%s: for %s\n", __func__, eulabeia_aggregate_to_str(aggregate));
+	fprintf(stderr,
+		"%s: for %s\n",
+		__func__,
+		eulabeia_aggregate_to_str(aggregate));
 
 	for (i = 0; i < toc->len; i++) {
 		type = toc->types[i];
